+++ lang/src/org/partiql/lang/planner/PartiQLPlannerDefault.kt
@@ -21,10 +21,8 @@ import org.partiql.lang.domains.PartiqlPhysical
 import org.partiql.lang.errors.ProblemCollector
 import org.partiql.lang.eval.CompileOptions
 import org.partiql.lang.eval.TypedOpBehavior
+import org.partiql.lang.eval.visitors.AggregationVisitorTransform
 import org.partiql.lang.eval.visitors.FromSourceAliasVisitorTransform
-import org.partiql.lang.eval.visitors.GroupByItemAliasVisitorTransform
-import org.partiql.lang.eval.visitors.GroupByPathExpressionVisitorTransform
-import org.partiql.lang.eval.visitors.GroupKeyReferencesVisitorTransform
 import org.partiql.lang.eval.visitors.OrderBySortSpecVisitorTransform
 import org.partiql.lang.eval.visitors.PartiqlAstSanityValidator
 import org.partiql.lang.eval.visitors.PipelinedVisitorTransform
@@ -116,9 +114,7 @@ internal class PartiQLPlannerDefault(
             SelectListItemAliasVisitorTransform(),
             FromSourceAliasVisitorTransform(),
             OrderBySortSpecVisitorTransform(),
-            GroupByItemAliasVisitorTransform(),
-            GroupByPathExpressionVisitorTransform(),
-            GroupKeyReferencesVisitorTransform(),
+            AggregationVisitorTransform(),
             SelectStarVisitorTransform()
         )
         return transform.transformStatement(this)